Fixed issue with model and collection generation script that wasn't stopping function name collisions (caused by two functions with the same name due to multiple tables referring back to it).
Also updated all models and collections to use order_by/offset/limit (when appropriate, eg, one to many relationships).

Fixed missing link_model getters that were not there for some reason, not sure when they vanished (not that it's very common to use link models).

Collection preloaders now require the $_var key to be provided, no longer as just an "override", this will make it more obvious what the function call is doing.